pyproject.toml
[tool.poetry]
name = "workflow_api"
version = "0.1.0"
description = ""
authors = ["Daryna Rodicheva <rdchvd29@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
networkx = "^3.2.1"
SQLAlchemy = "^2.0.29"
SQLAlchemy-Utils = "^0.41.2"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
fastapi = "^0.110.1"
fastapi-utilities = "^0.2.0"
PyJWT = "^2.8.0"
uvicorn = "^0.29.0"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"

[tool.poetry.dev-dependencies]
ruff = "^0.3.5"
pre-commit = "^3.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "Q", "I", "N"]

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"
